Moved EntityLink to its own type
