+++ src/pocketmine/network/mcpe/protocol/DataPacket.php
@@ -30,6 +30,7 @@ use pocketmine\entity\Entity;
 use pocketmine\item\ItemFactory;
 use pocketmine\math\Vector3;
 use pocketmine\network\mcpe\NetworkSession;
+use pocketmine\network\mcpe\protocol\types\EntityLink;
 use pocketmine\utils\BinaryStream;
 use pocketmine\utils\Utils;
 
@@ -514,20 +515,26 @@ abstract class DataPacket extends BinaryStream{
 	}
 
 	/**
-	 * @return array
+	 * @return EntityLink
 	 */
-	protected function getEntityLink() : array{
-		return [$this->getEntityUniqueId(), $this->getEntityUniqueId(), $this->getByte(), $this->getByte()];
+	protected function getEntityLink() : EntityLink{
+		$link = new EntityLink();
+		$link->fromEntityUniqueId = $this->getEntityUniqueId();
+		$link->toEntityUniqueId = $this->getEntityUniqueId();
+		$link->type = $this->getByte();
+		$link->byte2 = $this->getByte();
+
+		return $link;
 	}
 
 	/**
-	 * @param array $link
+	 * @param EntityLink $link
 	 */
-	protected function putEntityLink(array $link){
-		$this->putEntityUniqueId($link[0]);
-		$this->putEntityUniqueId($link[1]);
-		$this->putByte($link[2]);
-		$this->putByte($link[3]);
+	protected function putEntityLink(EntityLink $link){
+		$this->putEntityUniqueId($link->fromEntityUniqueId);
+		$this->putEntityUniqueId($link->toEntityUniqueId);
+		$this->putByte($link->type);
+		$this->putByte($link->byte2);
 
 	}
 }
